backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
psycopg[binary]==3.2.10
pydantic-settings==2.10.1
python-json-logger==3.3.0
numpy==2.3.2
pandas==2.3.1
scikit-learn==1.7.1
pytest==8.4.1
httpx==0.28.1
